+++ bitreader/include/bitreader/data_source/memory_byte_source.hpp
@@ -13,7 +13,7 @@ namespace brcpp
         memory_byte_source(const uint8_t* data, size_t size);
         size_t get_n(uint64_t& buf, size_t bytes);
         bool depleted();
-        uint64_t available();
+        size_t available();
         uint64_t position();
         void seek(uint64_t position);
         void skip(uint64_t bytes);
